Fix bug: owl selector selects only child of <body>
BUG: `body * + *` does not exclude additional top margin on top of what
is intended to be present because our `browser-sync` utility inserts a
<script> tag inside and at the top of <body>. Therefore our only child
of <body> becomes the second child and is selected by `* + *` or even
`body * + *`.
FIX: Add wrapper inside body and scope owl selector to
it.  But as a side effect we increase specificity of the ruleset.

LESSON:
Beware of (<script>) tags inserted by auxiliary utilities /
browser extensions that mess up w/ the order of source tags, e.g. making
first child a second child.