Size fexbox items w/ flex basis
Specify flex basis for each item. This way when we exceed 100% rely on
flex shrink.
NOTE: In the end we get a narrower main area (and wider sidebar) as
compared with the previous version which relied on flex grow factor and
flex basis 0%.
Reason - the item (column-main.tile) has padding (while sidebar does
not) which affected initial main size calculation when flex-bases was
0%.
LESSON: Providing specific 'main-size' in float-basis gives more precise
result (as compared to flex basis 0% and grow factor).